+++ AEDA_Airports/Plane.h
@@ -20,10 +20,10 @@ private:
 	string type; //para simplificar apenas existem modelos A, B, C.
 	int capacity;
 	vector<Flight*> flights; //não esquecer de usar new
-	float expenses;
+	double expenses;
 public:
 	Plane(string type, int capacity, vector<Flight*> &flights);
-	float calcExp(string type, vector<Flight*> &flights) const;
+	double calcExp(string type, vector<Flight*> &flights) const;
 	string getType() const;
 	int getCapacity() const;
 	vector<Flight*> getFlights() const;
